+++ modules/default/calendar/debug.js
@@ -15,7 +15,7 @@ var maximumEntries = 10;
 var maximumNumberOfDays = 365;
 var user = "magicmirror";
 var pass = "MyStrongPass";
-var includePastEvents = false;
+var broadcastPastEvents = false;
 
 var auth = {
 	user: user,